dropped pass-throughs and program accountinfos
