+++ solana/src/cpi/mod.rs
@@ -1,4 +1,4 @@
 pub mod associated_token;
-pub mod nautilus;
 pub mod system;
 pub mod token;
+pub mod token_metadata;
