+++ solana/src/objects/record/index.rs
@@ -133,14 +133,12 @@ impl<'a> NautilusIndex<'a> {
         &mut self,
         table_name: &str,
         fee_payer: impl NautilusSigner<'a>,
-        system_program: Box<AccountInfo<'a>>,
     ) -> Result<u32, ProgramError> {
         let count = self.data.add_record(table_name);
         cpi::system::transfer(
             fee_payer,
             Mut::<Self>::new(self.clone()),
             self.required_rent()? - self.lamports(),
-            system_program,
         )?;
         self.account_info.realloc(self.span()?, true)?;
         self.data
@@ -224,7 +222,7 @@ impl<'a> NautilusCreate<'a> for Create<'a, NautilusIndex<'a>> {
             index: std::collections::HashMap::new(),
         };
         let data_pointer = Box::new(data);
-        cpi::nautilus::create_record(
+        cpi::system::create_record(
             self.self_account.clone(),
             self.self_account.program_id,
             payer,
@@ -240,7 +238,7 @@ impl<'a> NautilusCreate<'a> for Create<'a, NautilusIndex<'a>> {
             index: std::collections::HashMap::new(),
         };
         let data_pointer = Box::new(data);
-        cpi::nautilus::create_record(
+        cpi::system::create_record(
             self.self_account.clone(),
             self.self_account.program_id,
             payer,
