+++ solana/src/objects/token/mint.rs
@@ -110,15 +110,14 @@ impl<'a> NautilusCreateMint<'a> for Create<'a, Mint<'a>> {
             account_info: self.fee_payer.to_owned(),
             system_program: self.system_program.to_owned(),
         });
-        cpi::nautilus::create_mint(
+        cpi::system::create_account(self.clone(), self.self_account.token_program.key, payer)?;
+        cpi::token::initialize_mint(
+            self.self_account.token_program.key,
             self.clone(),
+            mint_authority.key(),
+            freeze_authority.map(|f| f.key()),
             decimals,
-            mint_authority,
-            freeze_authority,
-            payer,
             self.rent.to_owned(),
-            self.system_program.to_owned(),
-            self.self_account.token_program.to_owned(),
         )?;
         self.self_account = Mint::load(
             self.self_account.account_info.clone(),
@@ -134,15 +133,14 @@ impl<'a> NautilusCreateMint<'a> for Create<'a, Mint<'a>> {
         freeze_authority: Option<impl NautilusAccountInfo<'a>>,
         payer: impl NautilusSigner<'a>,
     ) -> ProgramResult {
-        cpi::nautilus::create_mint(
+        cpi::system::create_account(self.clone(), self.self_account.token_program.key, payer)?;
+        cpi::token::initialize_mint(
+            self.self_account.token_program.key,
             self.clone(),
+            mint_authority.key(),
+            freeze_authority.map(|f| f.key()),
             decimals,
-            mint_authority,
-            freeze_authority,
-            payer,
             self.rent.to_owned(),
-            self.system_program.to_owned(),
-            self.self_account.token_program.to_owned(),
         )?;
         self.self_account = Mint::load(
             self.self_account.account_info.clone(),
